Fixe error in 2log device groups
An itemInserted signal was triggered in the backend, although
an append was originally called in the request handler. This
caused the clients to receive both an append and an insert
message for each new group. This bug had no worse consequences,
as the insert directly overwrote the newly created group.